refactor: VNC connection configs to VNC module
