+++ client/src/vnc/mod.rs
@@ -1,4 +1,7 @@
+#![allow(unused)]
+
 mod connect;
 mod auth;
 
 pub use self::connect::connect;
+pub use self::connect::Connection;
