+++ emulator/src/main.rs
@@ -37,7 +37,7 @@ fn main() -> Result<(), Error> {
 
 fn local_config() -> Config<'static> {
     Config{
-        connection: einkvnc::config::Connection { host: "localhost", port: 5901, username: None, password: Some("123456"), exclusive: false },
+        connection: einkvnc::vnc::Connection { host: "localhost", port: 5901, username: None, password: Some("123456"), exclusive: false },
         processing: einkvnc::processing::PostProcConfig { contrast_exp: 1.1, contrast_gray_point: 224.0, white_cutoff: 225 },
         rotate: 1,
         view_only: true,
